requirements.txt
PySide6>=6.6.0
PyMuPDF>=1.23.0
pythonocc-core>=7.7.0
numpy>=1.24.0
Pillow>=10.0.0
